app/requirements.txt
python-dotenv~=1.0.0
pandas~=2.1.0
Faker~=19.6.0
SQLAlchemy~=2.0.20